Extract Octokit wrappers
Improve legibility of main file
