+++ index.js
@@ -1,5 +1,11 @@
 const core = require('@actions/core');
 const github = require('@actions/github');
+const {
+  deleteOrgContainerVersion,
+  deleteUserContainerVersion,
+  listOrgContainerVersions,
+  listUserContainerVersions,
+} = require('./src/octokit');
 const {getPruningList} = require('./src/pruning');
 const {versionFilter} = require('./src/version-filter');
 
@@ -10,36 +16,6 @@ const versionSummary = (version) => ({
   name: version.name,
   created_at: version.created_at,
   tags: version.metadata.container.tags,
-});
-
-const listOrgContainerVersions = (octokit) => (organization, container) => (pageSize, page = 1) => octokit.rest.packages.getAllPackageVersionsForPackageOwnedByOrg({
-  package_type: 'container',
-  org: organization,
-  package_name: container,
-  page,
-  per_page: pageSize,
-  state: 'active',
-});
-
-const listUserContainerVersions = (octokit) => (container) => (pageSize, page = 1) => octokit.rest.packages.getAllPackageVersionsForPackageOwnedByAuthenticatedUser({
-  package_type: 'container',
-  package_name: container,
-  page,
-  per_page: pageSize,
-  state: 'active',
-});
-
-const deleteOrgContainerVersion = (octokit) => (organization, container) => (version) => octokit.rest.packages.deletePackageVersionForOrg({
-  package_type: 'container',
-  org: organization,
-  package_name: container,
-  package_version_id: version.id,
-});
-
-const deleteUserContainerVersion = (octokit) => (container) => (version) => octokit.rest.packages.deletePackageVersionForAuthenticatedUser({
-  package_type: 'container',
-  package_name: container,
-  package_version_id: version.id,
 });
 
 const dryRunDelete = (version) => new Promise((resolve) => {
